+++ primitives/Cargo.toml
@@ -10,14 +10,14 @@ codec = { package = "parity-scale-codec", version = "3.0.0", default-features = 
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
 
-# substrate
-frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
-frame-system = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
-pallet-assets = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
-pallet-balances = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
+# substrate no_std
 sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
+# substrate std
+frame-system = { optional = true, git = "https://github.com/paritytech/substrate.git", branch = "master" }
+pallet-assets = { optional = true, git = "https://github.com/paritytech/substrate.git", branch = "master" }
+pallet-balances = { optional = true, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 
 [dev-dependencies]
 node-template-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "master" }
@@ -28,12 +28,12 @@ std = [
     "codec/std",
     "hex/std",
     "scale-info/std",
-    # substrate
-    "frame-support/std",
-    "frame-system/std",
-    "pallet-assets/std",
-    "pallet-balances/std",
+    # substrate no_std
     "sp-core/std",
     "sp-runtime/std",
     "sp-std/std",
+    # substrate std
+    "frame-system",
+    "pallet-assets",
+    "pallet-balances",
 ]
