+++ primitives/src/pallet_traits/mod.rs
@@ -1,0 +1,24 @@
+/*
+   Copyright 2019 Supercomputing Systems AG
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+	   http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+
+*/
+
+pub use frame_system_config::FrameSystemConfig;
+pub use pallet_assets_config::AssetsConfig;
+pub use pallet_balances_config::BalancesConfig;
+
+pub mod frame_system_config;
+pub mod pallet_assets_config;
+pub mod pallet_balances_config;
