+++ primitives/src/pallet_traits/pallet_balances_config.rs
@@ -1,0 +1,44 @@
+/*
+   Copyright 2019 Supercomputing Systems AG
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+	   http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+
+*/
+use codec::{Codec, MaxEncodedLen};
+use core::fmt::Debug;
+use scale_info::TypeInfo;
+use sp_runtime::{
+	traits::{AtLeast32BitUnsigned, Get, MaybeSerializeDeserialize},
+	FixedPointOperand,
+};
+/// Simplifed pallet balances Config trait. Needed because substrate pallets compile to wasm
+/// in no_std mode.
+pub trait BalancesConfig: crate::FrameSystemConfig {
+	type Balance: AtLeast32BitUnsigned
+		+ Codec
+		+ Default
+		+ Copy
+		+ MaybeSerializeDeserialize
+		+ Debug
+		+ MaxEncodedLen
+		+ TypeInfo
+		+ FixedPointOperand;
+	type DustRemoval;
+	type RuntimeEvent;
+	type ExistentialDeposit: Get<Self::Balance>;
+	type AccountStore;
+	type WeightInfo;
+	type MaxLocks: Get<u32>;
+	type MaxReserves: Get<u32>;
+	type ReserveIdentifier: MaxEncodedLen + Ord + Copy;
+}
